# [PERF] sale_stock: improve perf of _get_outgoing_incoming_moves