Issue -->

A noticeable amount of time is spent in the method `_get_outgoing_incoming_moves` when doing the `OR` operation on the stock.move recordsets when there is a large number of moves to iterate through.

Solution -->

Replace the odoo recordset operations with set operations.

Benchmarks -->

| # of moves | Before | After |
|--------|--------|--------|
| 32000 | 24.08s | 1.43s |
| 7300 | 1.74s | 0.40s | 

opw-4220302