@robodoo 2024-10-14
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/183671.png)](https://mergebot.odoo.com/odoo/odoo/pull/183671)