@ticodoo 2024-10-22
Hello, fix looks good to me, but can you fix up your commit message?

Please use a tag that matches the [standard](https://www.odoo.com/documentation/master/contributing/development/git_guidelines.html#commit-message-structure). In this case, it should be `[FIX]` and you should include some sort of benchmark value in your commit message (for an example: see https://github.com/odoo/odoo/pull/139067/commits/7db04acc2758c938e6df2075b7ae1cb3f060941c).

Usually we prefer a flamegraph (in the PR) for showing where the problematic method is, but this is a pretty clear case. I'd prefer your commit message to reflect that:

_Using the `OR` operation can be significantly slow when dealing with large recordsets. Therefore we use the faster set.add() on their ids and rely on the cache to quickly browse the needed records afterwards._