@adda-odoo 2024-10-23
Hi @ticodoo, I've copied over the benchmarks from the PR description to the commit message!

> In this case, it should be `[FIX]`

I believe that tag is reserved for bug fixes and not performance improvements (at least not anymore).