@ticodoo 2024-10-24
@adda-odoo oops, didn't know about the new tag. Thanks for the heads up that the documentation needs to be updated :brain: 

@robodoo r+